Move singer display into better place.
